+++ flask_nacos/__init__.py
@@ -2,19 +2,25 @@
 
 from .exceptions import (
     FlaskNacosError,
+    NacosClientError,
     NacosConfigError,
+    NacosDeregistrationError,
     NacosDiscoveryError,
     NacosRegistrationError,
+    NacosValidationError,
 )
 from .extension import FlaskNacos
 
-__version__ = "0.1.0"
+__version__ = "0.2.0"
 
 __all__ = [
     "FlaskNacos",
     "FlaskNacosError",
     "NacosConfigError",
+    "NacosClientError",
+    "NacosValidationError",
     "NacosRegistrationError",
+    "NacosDeregistrationError",
     "NacosDiscoveryError",
     "__version__",
 ]